WIP ProgressListener

+++ src/main/java/de/embl/cba/bdp2/progress/ProgressListener.java
@@ -2,5 +2,5 @@ package de.embl.cba.bdp2.progress;
 
 public interface ProgressListener
 {
-	void progress( int current, int total );
+	void progress( long current, long total );
 }
